test(app): Move test utils out of __tests__
